Update file for fake hardware simulation and add file for gazebo simulation.
